requirements.txt
Flask
Flask-WTF
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
WTForms
email-validator
pytest